+++ spring-cloud-gcp-examples/spring-cloud-gcp-integration-pubsub-example/spring-cloud-gcp-integration-pubsub-example-receiver/pom.xml
@@ -3,19 +3,15 @@
 		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<parent>
-		<artifactId>spring-cloud-gcp-spring-integration-pubsub-example</artifactId>
+		<artifactId>spring-cloud-gcp-integration-pubsub-example</artifactId>
 		<groupId>com.example</groupId>
 		<version>1.0.0.BUILD-SNAPSHOT</version>
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
 
-	<artifactId>spring-cloud-gcp-spring-integration-pubsub-example-receiver</artifactId>
+	<artifactId>spring-cloud-gcp-integration-pubsub-example-receiver</artifactId>
 
 	<dependencies>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-web</artifactId>
-		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-integration-gcp</artifactId>
