+++ spring-cloud-gcp-examples/spring-cloud-gcp-integration-pubsub-example/spring-cloud-gcp-integration-pubsub-example-sender/src/main/java/com/example/WebController.java
@@ -17,12 +17,6 @@
 package com.example;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cloud.gcp.pubsub.core.PubSubOperations;
-import org.springframework.context.annotation.Bean;
-import org.springframework.integration.annotation.MessagingGateway;
-import org.springframework.integration.annotation.ServiceActivator;
-import org.springframework.integration.gcp.pubsub.outbound.PubSubMessageHandler;
-import org.springframework.messaging.MessageHandler;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -35,15 +29,7 @@ import org.springframework.web.servlet.view.RedirectView;
 public class WebController {
 
 	@Autowired
-	private PubsubOutboundGateway messagingGateway;
-
-	@Bean
-	@ServiceActivator(inputChannel = "pubsubOutputChannel")
-	public MessageHandler messageSender(PubSubOperations pubsubTemplate) {
-		PubSubMessageHandler outboundAdapter =
-				new PubSubMessageHandler(pubsubTemplate, "exampleTopic");
-		return outboundAdapter;
-	}
+	private SenderApplication.PubsubOutboundGateway messagingGateway;
 
 	/**
 	 * Posts a message to a Google Cloud Pub/Sub topic, through Spring's messaging gateway, and
@@ -55,11 +41,5 @@ public class WebController {
 	public RedirectView postMessage(@RequestParam("message") String message) {
 		this.messagingGateway.sendToPubsub(message);
 		return new RedirectView("/");
-	}
-
-	@MessagingGateway(defaultRequestChannel = "pubsubOutputChannel")
-	public interface PubsubOutboundGateway {
-
-		void sendToPubsub(String text);
 	}
 }
