--- spring-cloud-gcp-examples/spring-cloud-gcp-spring-integration-pubsub-example/spring-cloud-gcp-spring-integration-pubsub-example-receiver/src/main/java/com/example/ReceiverApplication.java
@@ -1,30 +1,0 @@
-/*
- *  Copyright 2017 original author or authors.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package com.example;
-
-import java.io.IOException;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ReceiverApplication {
-
-	public static void main(String[] args) throws IOException {
-		SpringApplication.run(ReceiverApplication.class, args);
-	}
-}
